SPIAT: Spatial Image Analysis of Tissues

SPIAT (**Sp**atial **I**mage **A**nalysis of **T**issues) is an R package with a suite of data processing, quality control, visualization and data analysis tools. SPIAT is compatible with data generated from single-cell spatial proteomics platforms (e.g. OPAL, CODEX, MIBI, cellprofiler). SPIAT reads spatial data in the form of X and Y coordinates of cells, marker intensities and cell phenotypes. SPIAT includes six analysis modules that allow visualization, calculation of cell colocalization, categorization of the immune microenvironment relative to tumor areas, analysis of cellular neighborhoods, and the quantification of spatial heterogeneity, providing a comprehensive toolkit for spatial data analysis.

See also

Author

Maintainer: Yuzhou Feng yuzhou.feng@petermac.org (ORCID)

Authors:

Anna Trigos anna.trigos@petermac.org (ORCID)

Tianpei Yang yangtp3344@gmail.com

Mabel Li

John Zhu

Volkan Ozcoban volkanozcoban1@gmail.com

Maria Doyle maria.doyle@petermac.org