Calculates Ripley's K-function with border correction

Returns K function of cell distribution with options to specify cell phenotype, plot, plot confidence envelope and plot confidence interval of K function. Plotted K-functions utilise an inhomogeneous Poisson distribution with point intensity produced to reflect the original data as the null distribution. Unable to plot both confidence interval and envelope simultaneously.

Usage

calculate_k_function(
  point_pattern,
  phenotypes_of_interest = NULL,
  determine_confidence = NULL,
  plot_envelope = NULL,
  plot = 1
)

Arguments

point_pattern

ppp object (marked point pattern).

phenotypes_of_interest

String supplying the phenotype of interest, if unspecified the entire point pattern will be considered

determine_confidence

Indicate to plot confidence interval of the K function calculated (adds significant run time)

plot_envelope

Indicates whether the significance envelope of simulated data should be plotted (based on inhomogeneous Poisson distribution)

plot

Indicates whether to plot K-function

Details

determine_confidence: assign any integer value to indicate to plot the confidence interval for the K-function itself. Set to NULL as default, calculating this can be computationally expensive phenotypes_of_interest: the cell phenotype/phenotypes to be considered in the K-function calculation if not the entire ppp object. Set to NULL as default plot_envelope: assign any integer value to indicate to plot the envelope of the simulated inhomogeneous poisson distribution plot: assign any integer value to indicate to plot the K-function itself, default is to plot. Set to NULL if plot is not to be produced Outputs - k_function: Array containing values required to plot K function and the associated inhomogeneous Poisson distribution (null distribution)