calculate_pairwise_distances_between_celltypes

Returns the pairwise distances between cells of different types. If none of the cell types are found, it will print an error message and return a vector of NAs.

Usage

calculate_pairwise_distances_between_celltypes(
  sce_object,
  cell_types_of_interest = NULL,
  feature_colname
)

Arguments

sce_object

SingleCellExperiment object in the form of the output of format_image_to_sce.

cell_types_of_interest

String Vector containing cell types to be considered, if NULL, all cell type combinations will be calculated.

feature_colname

String of the name the feature column with the cell types of interest to be considered.

Value

A data.frame is returned.

Examples

calculate_pairwise_distances_between_celltypes(SPIAT::defined_image, 
cell_types_of_interest = c("Tumour","Immune1"), feature_colname = "Cell.Type")
#>         Cell1   Cell2   Distance            Pair
#> 2     Cell_25 Cell_15  119.39827 Immune1_Immune1
#> 3     Cell_30 Cell_15   61.77859 Immune1_Immune1
#> 4     Cell_48 Cell_15  279.84944 Immune1_Immune1
#> 5     Cell_56 Cell_15  237.67335 Immune1_Immune1
#> 6     Cell_61 Cell_15  335.22489 Immune1_Immune1
#> 7     Cell_71 Cell_15  342.60619 Immune1_Immune1
#> 8     Cell_83 Cell_15  301.84395 Immune1_Immune1
#> 9     Cell_86 Cell_15  225.22881 Immune1_Immune1
#> 10    Cell_94 Cell_15  542.32855 Immune1_Immune1
#> 11    Cell_99 Cell_15  405.84255 Immune1_Immune1
#> 12   Cell_105 Cell_15  514.01023 Immune1_Immune1
#> 13   Cell_107 Cell_15  435.39893 Immune1_Immune1
#> 14   Cell_154 Cell_15  706.99725 Immune1_Immune1
#> 15   Cell_155 Cell_15  608.28345 Immune1_Immune1
#> 16   Cell_162 Cell_15  591.03913 Immune1_Immune1
#> 17   Cell_181 Cell_15  911.64023 Immune1_Immune1
#> 18   Cell_195 Cell_15  866.76045 Immune1_Immune1
#> 19   Cell_196 Cell_15  961.34659 Immune1_Immune1
#> 20   Cell_205 Cell_15 1047.02696 Immune1_Immune1
#> 21   Cell_211 Cell_15 1034.07460 Immune1_Immune1
#> 22   Cell_227 Cell_15 1102.73314 Immune1_Immune1
#> 23   Cell_242 Cell_15  984.53051 Immune1_Immune1
#> 24   Cell_244 Cell_15 1026.58735 Immune1_Immune1
#> 25   Cell_251 Cell_15 1250.27644 Immune1_Immune1
#> 26   Cell_252 Cell_15 1230.93722 Immune1_Immune1
#> 27   Cell_262 Cell_15 1343.33578 Immune1_Immune1
#> 28   Cell_271 Cell_15 1394.29934 Immune1_Immune1
#> 29   Cell_284 Cell_15 1291.47258 Immune1_Immune1
#> 30   Cell_309 Cell_15 1484.68671 Immune1_Immune1
#> 31   Cell_310 Cell_15 1461.57290 Immune1_Immune1
#> 32   Cell_321 Cell_15 1475.60228 Immune1_Immune1
#> 33   Cell_324 Cell_15 1550.22397 Immune1_Immune1
#> 34   Cell_326 Cell_15 1533.05486 Immune1_Immune1
#> 35   Cell_331 Cell_15 1552.32839 Immune1_Immune1
#> 36   Cell_333 Cell_15 1441.70710 Immune1_Immune1
#> 37   Cell_336 Cell_15 1457.42721 Immune1_Immune1
#> 38   Cell_352 Cell_15 1597.35446 Immune1_Immune1
#> 39   Cell_366 Cell_15 1645.71833 Immune1_Immune1
#> 40   Cell_410 Cell_15 1970.70975 Immune1_Immune1
#> 41   Cell_432 Cell_15 1886.88350 Immune1_Immune1
#> 42   Cell_464 Cell_15  106.16733 Immune1_Immune1
#> 43   Cell_465 Cell_15  191.51914 Immune1_Immune1
#> 44   Cell_466 Cell_15  196.15280 Immune1_Immune1
#> 45   Cell_469 Cell_15  186.99581 Immune1_Immune1
#> 46   Cell_486 Cell_15  386.12403 Immune1_Immune1
#> 47   Cell_488 Cell_15  217.28974 Immune1_Immune1
#> 48   Cell_490 Cell_15  350.11600 Immune1_Immune1
#> 49   Cell_492 Cell_15  362.94678 Immune1_Immune1
#> 50   Cell_497 Cell_15  380.74370 Immune1_Immune1
#> 51   Cell_498 Cell_15  310.68063 Immune1_Immune1
#> 52   Cell_505 Cell_15  206.15130 Immune1_Immune1
#> 53   Cell_533 Cell_15  347.57692 Immune1_Immune1
#> 54   Cell_563 Cell_15  470.33985 Immune1_Immune1
#> 55   Cell_568 Cell_15  481.44486 Immune1_Immune1
#> 56   Cell_574 Cell_15  474.50684 Immune1_Immune1
#> 57   Cell_613 Cell_15  679.92114 Immune1_Immune1
#> 58   Cell_622 Cell_15  787.19320 Immune1_Immune1
#> 59   Cell_626 Cell_15  663.90680 Immune1_Immune1
#> 60   Cell_645 Cell_15  965.06351 Immune1_Immune1
#> 61   Cell_649 Cell_15  968.63505 Immune1_Immune1
#> 62   Cell_654 Cell_15  912.29932 Immune1_Immune1
#> 63   Cell_663 Cell_15  942.61238 Immune1_Immune1
#> 64   Cell_669 Cell_15  991.26742 Immune1_Immune1
#> 65   Cell_703 Cell_15 1038.46534 Immune1_Immune1
#> 66   Cell_706 Cell_15 1023.87276 Immune1_Immune1
#> 67   Cell_710 Cell_15 1023.22733 Immune1_Immune1
#> 68   Cell_742 Cell_15 1053.17969 Immune1_Immune1
#> 69   Cell_748 Cell_15 1274.64357 Immune1_Immune1
#> 70   Cell_817 Cell_15 1589.61641 Immune1_Immune1
#> 71   Cell_835 Cell_15 1500.45282 Immune1_Immune1
#> 72   Cell_847 Cell_15 1681.61651 Immune1_Immune1
#> 73   Cell_891 Cell_15 1737.45361 Immune1_Immune1
#> 74   Cell_896 Cell_15 1720.25833 Immune1_Immune1
#> 75   Cell_907 Cell_15 1811.04895 Immune1_Immune1
#> 76   Cell_910 Cell_15 1887.10633 Immune1_Immune1
#> 77   Cell_926 Cell_15 1986.01555 Immune1_Immune1
#> 78   Cell_952 Cell_15  397.79650 Immune1_Immune1
#> 79   Cell_959 Cell_15  504.82750 Immune1_Immune1
#> 80  Cell_1004 Cell_15  436.51426 Immune1_Immune1
#> 81  Cell_1006 Cell_15  357.79435 Immune1_Immune1
#> 82  Cell_1015 Cell_15  452.95784 Immune1_Immune1
#> 83  Cell_1018 Cell_15  503.36362 Immune1_Immune1
#> 84  Cell_1021 Cell_15  527.32357 Immune1_Immune1
#> 85  Cell_1027 Cell_15  425.10964 Immune1_Immune1
#> 86  Cell_1036 Cell_15  467.69452 Immune1_Immune1
#> 87  Cell_1067 Cell_15  628.78794 Immune1_Immune1
#> 88  Cell_1104 Cell_15  716.37391 Immune1_Immune1
#> 89  Cell_1108 Cell_15  712.62464 Immune1_Immune1
#> 90  Cell_1114 Cell_15  783.66738 Immune1_Immune1
#> 91  Cell_1136 Cell_15 1011.86062 Immune1_Immune1
#> 92  Cell_1389 Cell_15 1935.00093 Immune1_Immune1
#> 93  Cell_1393 Cell_15 1939.40134 Immune1_Immune1
#> 94  Cell_1396 Cell_15 1967.53615 Immune1_Immune1
#> 95  Cell_1402 Cell_15 1895.27981 Immune1_Immune1
#> 96  Cell_1405 Cell_15 1888.26272 Immune1_Immune1
#> 97  Cell_1412 Cell_15 1948.32303 Immune1_Immune1
#> 98  Cell_1426 Cell_15 1876.60613 Immune1_Immune1
#> 99  Cell_1431 Cell_15 1900.83173 Immune1_Immune1
#> 100 Cell_1455 Cell_15  552.73287 Immune1_Immune1
#> 101 Cell_1463 Cell_15  673.67090 Immune1_Immune1
#> 102 Cell_1465 Cell_15  583.78598 Immune1_Immune1
#> 103 Cell_1467 Cell_15  528.64096 Immune1_Immune1
#> 104 Cell_1475 Cell_15  516.95562 Immune1_Immune1
#> 105 Cell_1479 Cell_15  539.53320 Immune1_Immune1
#> 106 Cell_1490 Cell_15  634.10341 Immune1_Immune1
#> 107 Cell_1515 Cell_15  522.80830 Immune1_Immune1
#> 108 Cell_1524 Cell_15  709.77551 Immune1_Immune1
#> 109 Cell_1688 Cell_15 1146.20574 Immune1_Immune1
#> 110 Cell_1721 Cell_15 1290.72533 Immune1_Immune1
#> 111 Cell_1722 Cell_15 1146.21800 Immune1_Immune1
#> 112 Cell_1888 Cell_15 1879.38698 Immune1_Immune1
#> 113 Cell_1899 Cell_15 1941.01824 Immune1_Immune1
#> 114 Cell_1908 Cell_15 1883.91277 Immune1_Immune1
#> 115 Cell_1912 Cell_15 1932.52669 Immune1_Immune1
#> 116 Cell_1929 Cell_15 1931.73884 Immune1_Immune1
#> 117 Cell_1937 Cell_15 2033.07503 Immune1_Immune1
#> 118 Cell_1939 Cell_15 2039.46600 Immune1_Immune1
#> 119 Cell_1947 Cell_15  892.23909 Immune1_Immune1
#> 120 Cell_1950 Cell_15  796.89490 Immune1_Immune1
#> 121 Cell_1956 Cell_15  815.34079 Immune1_Immune1
#> 122 Cell_1974 Cell_15  792.70396 Immune1_Immune1
#> 123 Cell_1987 Cell_15  811.99958 Immune1_Immune1
#> 124 Cell_2006 Cell_15  727.99883 Immune1_Immune1
#> 125 Cell_2039 Cell_15  779.15222 Immune1_Immune1
#> 126 Cell_2208 Cell_15 1397.86105 Immune1_Immune1
#> 127 Cell_2220 Cell_15 1386.99855 Immune1_Immune1
#> 128 Cell_2243 Cell_15 1446.50924 Immune1_Immune1
#> 129 Cell_2256 Cell_15 1530.24110 Immune1_Immune1
#> 130 Cell_2259 Cell_15 1612.99515 Immune1_Immune1
#> 131 Cell_2262 Cell_15 1498.37579 Immune1_Immune1
#> 132 Cell_2273 Cell_15 1486.91357 Immune1_Immune1
#> 133 Cell_2278 Cell_15 1430.61063 Immune1_Immune1
#> 134 Cell_2297 Cell_15 1703.35470 Immune1_Immune1
#> 135 Cell_2330 Cell_15 1704.61831 Immune1_Immune1
#> 136 Cell_2345 Cell_15 1764.13668 Immune1_Immune1
#> 137 Cell_2352 Cell_15 1779.75394 Immune1_Immune1
#> 138 Cell_2354 Cell_15 1759.20561 Immune1_Immune1
#> 139 Cell_2355 Cell_15 1872.52528 Immune1_Immune1
#> 140 Cell_2396 Cell_15 1964.10273 Immune1_Immune1
#> 141 Cell_2399 Cell_15 1976.22403 Immune1_Immune1
#> 142 Cell_2413 Cell_15 2081.99058 Immune1_Immune1
#> 143 Cell_2420 Cell_15 2020.83573 Immune1_Immune1
#> 144 Cell_2426 Cell_15 2075.31779 Immune1_Immune1
#> 145 Cell_2434 Cell_15 2005.29147 Immune1_Immune1
#> 146 Cell_2462 Cell_15 1037.89033 Immune1_Immune1
#> 147 Cell_2463 Cell_15  918.77474 Immune1_Immune1
#> 148 Cell_2480 Cell_15 1074.40346 Immune1_Immune1
#> 149 Cell_2499 Cell_15  987.44730 Immune1_Immune1
#> 150 Cell_2500 Cell_15  971.97735 Immune1_Immune1
#> 151 Cell_2504 Cell_15  996.65277 Immune1_Immune1
#> 152 Cell_2508 Cell_15 1005.45611 Immune1_Immune1
#> 153 Cell_2536 Cell_15 1022.66563 Immune1_Immune1
#> 154 Cell_2567 Cell_15 1085.27977 Immune1_Immune1
#> 155 Cell_2569 Cell_15 1071.89290 Immune1_Immune1
#> 156 Cell_2582 Cell_15 1208.07205 Immune1_Immune1
#> 157 Cell_2778 Cell_15 1521.20994 Immune1_Immune1
#> 158 Cell_2779 Cell_15 1639.34594 Immune1_Immune1
#> 159 Cell_2793 Cell_15 1638.74555 Immune1_Immune1
#> 160 Cell_2825 Cell_15 1576.37749 Immune1_Immune1
#> 161 Cell_2830 Cell_15 1852.05740 Immune1_Immune1
#> 162 Cell_2834 Cell_15 1816.49965 Immune1_Immune1
#> 163 Cell_2835 Cell_15 1777.81124 Immune1_Immune1
#> 164 Cell_2858 Cell_15 1793.72599 Immune1_Immune1
#> 165 Cell_2882 Cell_15 2052.25334 Immune1_Immune1
#> 166 Cell_2895 Cell_15 1946.54621 Immune1_Immune1
#> 167 Cell_2916 Cell_15 1948.64476 Immune1_Immune1
#> 168 Cell_2925 Cell_15 2030.21670 Immune1_Immune1
#> 169 Cell_2936 Cell_15 2153.62013 Immune1_Immune1
#> 170 Cell_2965 Cell_15 2217.46012 Immune1_Immune1
#> 171 Cell_2986 Cell_15 1251.07717 Immune1_Immune1
#> 172 Cell_2998 Cell_15 1189.55029 Immune1_Immune1
#> 173 Cell_2999 Cell_15 1291.20140 Immune1_Immune1
#> 174 Cell_3011 Cell_15 1181.03533 Immune1_Immune1
#> 175 Cell_3017 Cell_15 1219.83878 Immune1_Immune1
#> 176 Cell_3022 Cell_15 1228.29514 Immune1_Immune1
#> 177 Cell_3035 Cell_15 1254.01220 Immune1_Immune1
#> 178 Cell_3051 Cell_15 1194.40956 Immune1_Immune1
#> 179 Cell_3053 Cell_15 1259.40230 Immune1_Immune1
#> 180 Cell_3057 Cell_15 1137.55045 Immune1_Immune1
#> 181 Cell_3086 Cell_15 1285.68956 Immune1_Immune1
#> 182 Cell_3098 Cell_15 1223.29567 Immune1_Immune1
#> 183 Cell_3115 Cell_15 1215.17120 Immune1_Immune1
#> 184 Cell_3122 Cell_15 1274.72351 Immune1_Immune1
#> 185 Cell_3133 Cell_15 1370.67796 Immune1_Immune1
#> 186 Cell_3306 Cell_15 1780.60681 Immune1_Immune1
#> 187 Cell_3335 Cell_15 1992.54465 Immune1_Immune1
#> 188 Cell_3336 Cell_15 1916.10567 Immune1_Immune1
#> 189 Cell_3349 Cell_15 1985.01477 Immune1_Immune1
#> 190 Cell_3353 Cell_15 1896.64365 Immune1_Immune1
#> 191 Cell_3376 Cell_15 2051.79296 Immune1_Immune1
#> 192 Cell_3379 Cell_15 2137.19287 Immune1_Immune1
#> 193 Cell_3381 Cell_15 2021.25806 Immune1_Immune1
#> 194 Cell_3385 Cell_15 2064.70304 Immune1_Immune1
#> 195 Cell_3399 Cell_15 2023.59299 Immune1_Immune1
#> 196 Cell_3421 Cell_15 2077.69702 Immune1_Immune1
#> 197 Cell_3425 Cell_15 2003.41968 Immune1_Immune1
#> 198 Cell_3431 Cell_15 2163.50939 Immune1_Immune1
#> 199 Cell_3471 Cell_15 2136.17199 Immune1_Immune1
#> 200 Cell_3482 Cell_15 2206.56098 Immune1_Immune1
#> 201 Cell_3494 Cell_15 1478.43441 Immune1_Immune1
#> 202 Cell_3496 Cell_15 1497.24636 Immune1_Immune1
#> 203 Cell_3511 Cell_15 1377.05245 Immune1_Immune1
#> 204 Cell_3538 Cell_15 1476.70492 Immune1_Immune1
#> 205 Cell_3539 Cell_15 1402.16665 Immune1_Immune1
#> 206 Cell_3548 Cell_15 1527.13158 Immune1_Immune1
#> 207 Cell_3549 Cell_15 1478.87546 Immune1_Immune1
#> 208 Cell_3552 Cell_15 1451.87065 Immune1_Immune1
#> 209 Cell_3553 Cell_15 1406.70661 Immune1_Immune1
#> 210 Cell_3564 Cell_15 1353.03657 Immune1_Immune1
#> 211 Cell_3572 Cell_15 1385.11606 Immune1_Immune1
#> 212 Cell_3656 Cell_15 1567.01825 Immune1_Immune1
#> 213 Cell_3658 Cell_15 1531.75939 Immune1_Immune1
#> 214 Cell_3661 Cell_15 1583.25450 Immune1_Immune1
#> 215 Cell_3675 Cell_15 1531.71176 Immune1_Immune1
#> 216 Cell_3698 Cell_15 1646.97589 Immune1_Immune1
#> 217 Cell_3701 Cell_15 1670.12070 Immune1_Immune1
#> 218 Cell_3706 Cell_15 1574.21185 Immune1_Immune1
#> 219 Cell_3713 Cell_15 1590.85245 Immune1_Immune1
#> 220 Cell_3721 Cell_15 1537.34877 Immune1_Immune1
#> 221 Cell_3864 Cell_15 2053.79851 Immune1_Immune1
#> 222 Cell_3882 Cell_15 2047.01429 Immune1_Immune1
#> 223 Cell_3886 Cell_15 1981.11065 Immune1_Immune1
#> 224 Cell_3888 Cell_15 2072.26180 Immune1_Immune1
#> 225 Cell_3905 Cell_15 2088.93542 Immune1_Immune1
#> 226 Cell_3906 Cell_15 2085.28265 Immune1_Immune1
#> 227 Cell_3911 Cell_15 2239.53936 Immune1_Immune1
#> 228 Cell_3912 Cell_15 2223.38917 Immune1_Immune1
#> 229 Cell_3919 Cell_15 2206.70140 Immune1_Immune1
#> 230 Cell_3930 Cell_15 2297.03786 Immune1_Immune1
#> 231 Cell_3942 Cell_15 2201.06633 Immune1_Immune1
#> 232 Cell_3944 Cell_15 2214.02518 Immune1_Immune1
#> 233 Cell_3993 Cell_15 2330.17150 Immune1_Immune1
#> 234 Cell_3996 Cell_15 2351.44299 Immune1_Immune1
#> 235 Cell_4003 Cell_15 1687.94859 Immune1_Immune1
#> 236 Cell_4005 Cell_15 1677.62154 Immune1_Immune1
#> 237 Cell_4022 Cell_15 1537.23092 Immune1_Immune1
#> 238 Cell_4026 Cell_15 1686.00666 Immune1_Immune1
#> 239 Cell_4027 Cell_15 1515.68654 Immune1_Immune1
#> 240 Cell_4030 Cell_15 1592.13804 Immune1_Immune1
#> 241 Cell_4051 Cell_15 1529.31722 Immune1_Immune1
#> 242 Cell_4062 Cell_15 1634.91926 Immune1_Immune1
#> 243 Cell_4068 Cell_15 1549.50088 Immune1_Immune1
#> 244 Cell_4076 Cell_15 1716.53888 Immune1_Immune1
#> 245 Cell_4084 Cell_15 1635.95359 Immune1_Immune1
#> 246 Cell_4105 Cell_15 1704.79760 Immune1_Immune1
#> 247 Cell_4107 Cell_15 1647.96090 Immune1_Immune1
#> 248 Cell_4115 Cell_15 1709.33194 Immune1_Immune1
#> 249 Cell_4126 Cell_15 1745.43320 Immune1_Immune1
#> 250 Cell_4137 Cell_15 1698.37118 Immune1_Immune1
#> 251 Cell_4145 Cell_15 1792.04892 Immune1_Immune1
#>  [ reached 'max' / getOption("max.print") -- omitted 1337242 rows ]