check for initial underlying distribution (cluster or Homogenous Poisson)

check_distribution_types Takes in point pattern object and performs a quadrat test and clark evans test to determine the underlying pattern. The results and p value will be printed to console

check_distribution_type(
  point_pattern,
  quadrat_dim = c(3, 3),
  phenotypes_of_interest = NULL
)

Arguments

point_pattern
sample point pattern to be analysed

quadrat_dim
vector containing number of quadrants in x,y direction respectively

phenotypes_of_interest
String describing the phenotypes to be examined

Details

Input - point_pattern: ppp object representing image to be analysed quadrat_dim: Numeric vector containing the dimensions for image segmentation when performing the quadrat test. default is 3x3 segementation in x,y direction phenotypes_of_interest: String vector containing the names of the cell phenotypes to be examined in the case the entire point pattern is not relevant

Output - test_results: List containing 2 htest objects containing test data