marker_surface_plot

Generates a 3D surface plot of the level of the selected marker. Note that the image is blurred based on the 'num_splits' parameter.

marker_surface_plot(
  spe_object,
  num_splits,
  marker,
  x_position_min = NULL,
  x_position_max = NULL,
  y_position_min = NULL,
  y_position_max = NULL
)

Arguments

spe_object

SpatialExperiment object in the form of the output of format_image_to_spe.

num_splits

Integer specifying the number of splits on the image, higher splits equal to higher resolution. Recommendation: 10-100

marker

Marker to plot

x_position_min

Integer specifying the minimum x boundary to be splitted

x_position_max

Integer specifying the maximum x boundary to be splitted

y_position_min

Integer specifying the minimum y boundary to be splitted

y_position_max

Integer specifying the maximum y boundary to be splitted

Value

A plot is returned

Examples

marker_surface_plot(SPIAT::simulated_image, num_splits=15, marker="Immune_marker1")